fix: move possible checking of hyp3 jobs into service
This was to fix a bug with on demand button always showing up on mid-sized
baseline search views
