+++ src/app/app.module.ts
@@ -150,6 +150,7 @@ export const routes = [
     services.PairService,
     services.SceneSelectService,
     services.OnDemandService,
+    services.PossibleHyp3JobsService,
     {provide: SAVER, useFactory: getSaver},
     {
       provide: DateAdapter,
