+++ src/app/components/results-menu/results-menu.component.html
@@ -19,7 +19,6 @@
       <div class="resize-controls">
         <div mwlResizeHandle [resizeEdges]="{ top: !(imageViewerOpen$ | async) }" class="resize-handle-filler">
           <mat-icon svgIcon="hgrip" class="hgrip-icon"></mat-icon>
-
         </div>
       </div>
