+++ src/app/components/shared/selectors/dataset-selector/dataset-selector.component.html
@@ -26,7 +26,8 @@
                   <div class="item-title">
                     {{ dataset.name }}
                     <span class="dataset-light-text">
-                        <span *ngIf="dataset.id == 'SENTINEL-1'"matTooltip="Available for On Demand Processing (HyP3)">
+                        <span *ngIf="dataset.id == 'SENTINEL-1' || dataset.id == 'SENTINEL-1 BURSTS'"
+                          matTooltip="Available for On Demand Processing (HyP3)">
                           <mat-icon class="hyp3-icon" svgIcon="hyp3"></mat-icon>
                         </span>
                         <span *ngIf="dataset.beta"><em> (beta) </em></span>
